pyproject.toml
[tool.poetry]
name = "chatbot-project"
version = "0.1.0"
description = "A ChatGPT-powered chatbot using FastAPI"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.0"
openai = "^1.54.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"